+++ app/src/main/cpp/fire/util/shader.h
@@ -25,7 +25,7 @@ public:
 
     void uniform1f(const GLchar *name, GLfloat value);
 
-    void uniform3f(const GLchar *name, ivec3 vector);
+    void uniform3f(const GLchar *name, vec3 vector);
 private:
     GLuint createShader(GLenum type, const char* src);
